Applied project's auto formatter and removed obsolete imports.

+++ jsettlers.logic/src/jsettlers/logic/map/original/OriginalMapFileContent.java
@@ -14,6 +14,8 @@
  *******************************************************************************/
 package jsettlers.logic.map.original;
 
+import java.util.BitSet;
+
 import jsettlers.algorithms.partitions.IBlockingProvider;
 import jsettlers.algorithms.partitions.PartitionCalculatorAlgorithm;
 import jsettlers.common.landscape.ELandscapeType;
@@ -30,33 +32,30 @@ import jsettlers.logic.map.original.OriginalMapFileDataStructs.EMapResources;
 import jsettlers.logic.map.original.OriginalMapFileDataStructs.EMapSettlersType;
 import jsettlers.logic.map.original.OriginalMapFileDataStructs.EMapStackType;
 
-import java.util.BitSet;
-
-
 /**
  * @author Thomas Zeugner
  * @author codingberlin
  */
 public class OriginalMapFileContent implements IMapData {
 
-	//- Heigh of original maps are 0..225 and of remake 0..127
+	// - Heigh of original maps are 0..225 and of remake 0..127
 	private final static float ORIGINAL_TO_REMAKE_HEIGHT_FACTOR = 127f / 225f;
 	private final static float ORIGINAL_TO_REMAKE_RESOURCE_AMOUNT_FACTOR = 127f / 15f;
 
-	//--------------------------------------------------//
+	// --------------------------------------------------//
 	public static class MapPlayerInfo {
 		public int startX;
 		public int startY;
 		public String playerName;
 		public OriginalMapFileDataStructs.EMapNations nation;
-		
+
 		public MapPlayerInfo(int X, int Y, String playerName, int nationInt) {
 			this.startX = X;
 			this.startY = Y;
 			this.playerName = playerName;
 			this.nation = OriginalMapFileDataStructs.EMapNations.FromMapValue(nationInt);
 		}
-		
+
 		public MapPlayerInfo(int X, int Y) {
 			this.startX = X;
 			this.startY = Y;
@@ -64,36 +63,34 @@ public class OriginalMapFileContent implements IMapData {
 			this.nation = OriginalMapFileDataStructs.EMapNations.ROMANS;
 		}
 	}
-	
-	//--------------------------------------------------//
-	
-	
+
+	// --------------------------------------------------//
+
 	public int fileChecksum = 0;
-	
-	//- original maps are squared
+
+	// - original maps are squared
 	private int widthHeight = 0;
 	private int dataCount = 0;
 
-	private byte [] height = null;
+	private byte[] height = null;
 	private ELandscapeType[] landscapeType = null;
-	private MapObject [] mapObject = null ;
-	private byte [] accessible = null ;
-	private EResourceType [] resources = null;
-	private byte [] resourceAmount = null;
+	private MapObject[] mapObject = null;
+	private byte[] accessible = null;
+	private EResourceType[] resources = null;
+	private byte[] resourceAmount = null;
 	private short[] blockedPartitions = null;
-	
+
 	private MapPlayerInfo[] mapPlayerInfos;
-	
-	
+
 	public OriginalMapFileContent(int widthHeight) {
 		setWidthHeight(widthHeight);
 	}
-	
+
 	public void setWidthHeight(int widthHeight) {
 		this.widthHeight = widthHeight;
-		
+
 		dataCount = widthHeight * widthHeight;
-		
+
 		height = new byte[dataCount];
 		landscapeType = new ELandscapeType[dataCount];
 		mapObject = new MapObject[dataCount];
@@ -103,142 +100,145 @@ public class OriginalMapFileContent implements IMapData {
 		blockedPartitions = new short[dataCount];
 	}
 
-
 	public void setLandscapeHeight(int pos, int height) {
-		if ((pos<0) || (pos> dataCount)) return;
+		if ((pos < 0) || (pos > dataCount))
+			return;
 
-		//- apply scaling from original to remake
-		height = Math.round((float) height * ORIGINAL_TO_REMAKE_HEIGHT_FACTOR);
-		if (height < 0) height = 0;
-		
-		this.height[pos] = (byte)height;
+		// - apply scaling from original to remake
+		height = Math.round(height * ORIGINAL_TO_REMAKE_HEIGHT_FACTOR);
+		if (height < 0)
+			height = 0;
+
+		this.height[pos] = (byte) height;
 	}
 
-	
 	public void setLandscape(int pos, int type) {
-		if ((pos<0) || (pos> dataCount)) return;
+		if ((pos < 0) || (pos > dataCount))
+			return;
 
 		OriginalMapFileDataStructs.EOriginalLandscapeType originalType = OriginalMapFileDataStructs.EOriginalLandscapeType.getTypeByInt(type);
 
 		landscapeType[pos] = originalType.value;
 	}
-	
+
 	public void setMapObject(int pos, int type) {
-		if ((pos<0) || (pos> dataCount)) return;
+		if ((pos < 0) || (pos > dataCount))
+			return;
 
 		mapObject[pos] = OriginalMapFileDataStructs.EObjectType.getTypeByInt(type).getNewInstance();
 	}
-	
-	public void setPlayerCount(int count)
-	{
+
+	public void setPlayerCount(int count) {
 		mapPlayerInfos = new MapPlayerInfo[count];
-		
-		for (int i=0; i < count; i++) {
-			//- init new player with "random" start position
+
+		for (int i = 0; i < count; i++) {
+			// - init new player with "random" start position
 			mapPlayerInfos[i] = new MapPlayerInfo(20 + i * 10, 20 + i * 10);
 		}
 	}
-	
-	
-	public void setPlayer(int index, int x, int y, int NationType, String PlayerName)
-	{
-		//System.out.println("Player "+ Integer.toString(index) +" : "+ PlayerName +" @ ("+ x +" , "+ y +")");
-		
-		if ((index < 0) || (index >= mapPlayerInfos.length)) return;
-		
+
+	public void setPlayer(int index, int x, int y, int NationType, String PlayerName) {
+		// System.out.println("Player "+ Integer.toString(index) +" : "+ PlayerName +" @ ("+ x +" , "+ y +")");
+
+		if ((index < 0) || (index >= mapPlayerInfos.length))
+			return;
+
 		mapPlayerInfos[index].nation = OriginalMapFileDataStructs.EMapNations.FromMapValue(NationType);
 		mapPlayerInfos[index].startX = x;
 		mapPlayerInfos[index].startY = y;
 		mapPlayerInfos[index].playerName = PlayerName;
 	}
-	
-	public void setPlayer(int index, int x, int y, OriginalMapFileDataStructs.EMapNations NationType, String PlayerName)
-	{
-		if ((index < 0) || (index >= mapPlayerInfos.length)) return;
-		
+
+	public void setPlayer(int index, int x, int y, OriginalMapFileDataStructs.EMapNations NationType, String PlayerName) {
+		if ((index < 0) || (index >= mapPlayerInfos.length))
+			return;
+
 		mapPlayerInfos[index].nation = NationType;
 		mapPlayerInfos[index].startX = x;
 		mapPlayerInfos[index].startY = y;
 		mapPlayerInfos[index].playerName = PlayerName;
 	}
-	
-	
-	public void setMapObject(int x, int y, MapObject newMapObject)	{
-		
+
+	public void setMapObject(int x, int y, MapObject newMapObject) {
+
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return;
+
 		mapObject[pos] = newMapObject;
 	}
-	
-	public void setBuilding(int x, int y, int BType, int party, int countSword1, int countSword2, int countSword3, int countArcher1, int countArcher2, int countArcher3, int countSpear1, int countSpear2, int countSpear3) {
+
+	public void setBuilding(int x, int y, int BType, int party, int countSword1, int countSword2, int countSword3, int countArcher1, int countArcher2,
+			int countArcher3, int countSpear1, int countSpear2, int countSpear3) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return;
+
 		EMapBuildingType BuildingType = EMapBuildingType.getTypeByInt(BType);
-		
-		if (BuildingType == EMapBuildingType.NOT_A_BUILDING) return;
+
+		if (BuildingType == EMapBuildingType.NOT_A_BUILDING)
+			return;
 		if (BuildingType.value != null) {
-			mapObject[pos] = new BuildingObject(BuildingType.value, (byte)party);
+			mapObject[pos] = new BuildingObject(BuildingType.value, (byte) party);
 		}
 	}
-	
+
 	public void setSettler(int x, int y, int SType, int party) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return;
+
 		EMapSettlersType SettlerType = EMapSettlersType.getTypeByInt(SType);
-		
-		if (SettlerType == EMapSettlersType.NOT_A_SETTLER) return;
+
+		if (SettlerType == EMapSettlersType.NOT_A_SETTLER)
+			return;
 		if (SettlerType.value != null) {
-			mapObject[pos] = new MovableObject(SettlerType.value, (byte)party);
+			mapObject[pos] = new MovableObject(SettlerType.value, (byte) party);
 		}
 	}
-	
-	
-	
+
 	public void setStack(int x, int y, int SType, int count) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return;
+
 		EMapStackType StackType = EMapStackType.getTypeByInt(SType);
-		
-		if (StackType == EMapStackType.NOT_A_STACK) return;
+
+		if (StackType == EMapStackType.NOT_A_STACK)
+			return;
 		if (StackType.value != null) {
 			mapObject[pos] = new StackObject(StackType.value, count);
 		}
 	}
 
-	
-	
 	public void setAccessible(int pos, byte isAccessible) {
-		if ((pos < 0) || (pos >= dataCount)) return;
-		
+		if ((pos < 0) || (pos >= dataCount))
+			return;
+
 		accessible[pos] = isAccessible;
 	}
-	
+
 	public void setResources(int pos, int resourcesType, int resourcesAmount) {
-		if ((pos < 0) || (pos >= dataCount)) return;
+		if ((pos < 0) || (pos >= dataCount))
+			return;
 
 		EMapResources mapResources = EMapResources.getTypeByInt(resourcesType);
-		
+
 		if (resourcesAmount == 0) {
 			resources[pos] = EResourceType.NOTHING;
 			resourceAmount[pos] = 0;
 		} else {
 			resources[pos] = mapResources.value;
-			resourceAmount[pos] = (byte) Math.round(((float) resourcesAmount) * ORIGINAL_TO_REMAKE_RESOURCE_AMOUNT_FACTOR);
+			resourceAmount[pos] = (byte) Math.round((resourcesAmount) * ORIGINAL_TO_REMAKE_RESOURCE_AMOUNT_FACTOR);
 		}
 	}
 
-
-	//- free the Arrays
-	public void FreeBuffer()
-	{
+	// - free the Arrays
+	public void FreeBuffer() {
 		dataCount = 0;
 		height = null;
 		landscapeType = null;
@@ -248,17 +248,16 @@ public class OriginalMapFileContent implements IMapData {
 		resources = null;
 		resourceAmount = null;
 	}
-	
-	
-	//------------------------//
-	//-- Interface IMapData --//
-	//------------------------//
-	
+
+	// ------------------------//
+	// -- Interface IMapData --//
+	// ------------------------//
+
 	@Override
 	public int getWidth() {
 		return widthHeight;
 	}
-	
+
 	@Override
 	public int getHeight() {
 		return widthHeight;
@@ -267,29 +266,33 @@ public class OriginalMapFileContent implements IMapData {
 	@Override
 	public ELandscapeType getLandscape(int x, int y) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return ELandscapeType.WATER1;
-		
-		if (landscapeType[pos]==null) return ELandscapeType.GRASS;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return ELandscapeType.WATER1;
+
+		if (landscapeType[pos] == null)
+			return ELandscapeType.GRASS;
+
 		return landscapeType[pos];
 	}
-	
+
 	@Override
 	public MapObject getMapObject(int x, int y) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return null;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return null;
+
 		return mapObject[pos];
 	}
 
 	@Override
 	public byte getLandscapeHeight(int x, int y) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return 0;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return 0;
+
 		return height[pos];
 	}
 
@@ -299,54 +302,56 @@ public class OriginalMapFileContent implements IMapData {
 	@Override
 	public byte getResourceAmount(short x, short y) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return 0;
-		
-		if (resources[pos] == null) return 0;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return 0;
+
+		if (resources[pos] == null)
+			return 0;
+
 		return resourceAmount[pos];
 	}
 
 	@Override
 	public EResourceType getResourceType(short x, short y) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return EResourceType.NOTHING;
-		
-		if (resources[pos] == null) return EResourceType.NOTHING;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return EResourceType.NOTHING;
+
+		if (resources[pos] == null)
+			return EResourceType.NOTHING;
+
 		return resources[pos];
 	}
-	
+
 	/**
 	 * Gets the id of the blocked partition of the given position.
+	 * 
 	 * @return The id of the blocked partition the given position belongs to.
 	 */
 	@Override
 	public short getBlockedPartition(short x, short y) {
 		int pos = y * widthHeight + x;
-		
-		if ((pos < 0) || (pos >= dataCount)) return 0;
-		
+
+		if ((pos < 0) || (pos >= dataCount))
+			return 0;
+
 		return blockedPartitions[pos];
 	}
-	
-	
+
 	public void calculateBlockedPartitions() {
 		MilliStopWatch watch = new MilliStopWatch();
 
 		BitSet notBlockedSet = new BitSet(dataCount);
-		
-		for (int pos = 0; pos < dataCount; pos++)
-		{
+
+		for (int pos = 0; pos < dataCount; pos++) {
 			notBlockedSet.set(pos, !landscapeType[pos].isBlocking);
 		}
-		
 
 		PartitionCalculatorAlgorithm partitionCalculator = new PartitionCalculatorAlgorithm(0, 0, widthHeight, widthHeight, notBlockedSet,
 				IBlockingProvider.DEFAULT_IMPLEMENTATION);
 		partitionCalculator.calculatePartitions();
-
 
 		for (short y = 0; y < widthHeight; y++) {
 			for (short x = 0; x < widthHeight; x++) {
@@ -357,18 +362,16 @@ public class OriginalMapFileContent implements IMapData {
 		watch.stop("Calculating partitions needed");
 		System.out.println("found " + partitionCalculator.getNumberOfPartitions() + " partitions.");
 	}
-	
 
 	@Override
 	public ShortPoint2D getStartPoint(int player) {
-		if ((player < 0) || (player >= mapPlayerInfos.length))
-		{
-			System.out.print("Error: not a player for getStartPoint("+ player +")");
-			return new ShortPoint2D(100,100);
+		if ((player < 0) || (player >= mapPlayerInfos.length)) {
+			System.out.print("Error: not a player for getStartPoint(" + player + ")");
+			return new ShortPoint2D(100, 100);
 		}
 		return new ShortPoint2D(mapPlayerInfos[player].startX, mapPlayerInfos[player].startY);
 	}
-	
+
 	@Override
 	public int getPlayerCount() {
 		return mapPlayerInfos.length;
